+++ tests/extern_impl/types.h
@@ -1,0 +1,12 @@
+#ifndef INTERFACE99_TESTS_EXTERN_IMPL_TYPES_H
+#define INTERFACE99_TESTS_EXTERN_IMPL_TYPES_H
+
+typedef struct {
+    char dummy;
+} Foo1Impl;
+
+typedef struct {
+    char dummy;
+} Foo2Impl;
+
+#endif // INTERFACE99_TESTS_EXTERN_IMPL_TYPES_H