+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+ - Generate virtual tables with static linkage in `impl(Primary)` & `declImpl` [**BC**].
+
+### Added
+
+ - Counterparts with external linkage: `externImpl(Primary)` & `externDeclImpl`.
+
 ## [0.2.0] - 2021-07-01
 
 ### Added